Refractor discrete random variables to not all contain explicit pmfs
