+++ alea/discrete/root_randvar.py
@@ -3,13 +3,26 @@ from .discrete_randvar import DiscreteRandVar
 import random
 import math
 import numpy as np
+import copy
 
 
 class RootDiscreteRandVar(DiscreteRandVar):
 
     def __init__(self, sample_space, mass_function):
-        DiscreteRandVar.__init__(self, sample_space, mass_function)
-        self.elements = None
+        DiscreteRandVar.__init__(self)
+        self.sample_space = copy.copy(sample_space)
+        self.mass_function = mass_function
+        self.pcache = {}
+        self.sample_list = None
+
+
+    def probability_of(self, x):
+        if x in self.pcache:
+            return self.pcache[x]
+        else:
+            p = self.mass_function(x)
+            self.pcache[x] = p
+            return p
 
 
     def _new_sample(self):
@@ -17,10 +30,10 @@ class RootDiscreteRandVar(DiscreteRandVar):
         # variable's probability distribution. This, in turn, assumes 
         # that this random variable does not have any parents and thus
         # represents an independent, real-world event
-        if self.elements is None:
-            self.elements = list(self.sample_space)
-        probabilities = [self.probability_of(x) for x in self.elements]
-        return np.random.choice(self.elements, 1, p=probabilities)[0]
+        if self.sample_list is None:
+            self.sample_list = list(self.sample_space)
+        probabilities = [self.probability_of(x) for x in self.sample_list]
+        return np.random.choice(self.sample_list, 1, p=probabilities)[0]
 
 
     def _new_mean(self, fixed_means):
